Moved some deprecated methods to the corresponding files.
