+++ gdk/screen.go
@@ -74,18 +74,6 @@ func (v *Screen) GetSystemVisual() (*Visual, error) {
 	return visual, nil
 }
 
-// GetWidth is a wrapper around gdk_screen_get_width().
-func (v *Screen) GetWidth() int {
-	c := C.gdk_screen_get_width(v.native())
-	return int(c)
-}
-
-// GetHeight is a wrapper around gdk_screen_get_height().
-func (v *Screen) GetHeight() int {
-	c := C.gdk_screen_get_height(v.native())
-	return int(c)
-}
-
 // ScreenGetDefault is a wrapper aroud gdk_screen_get_default().
 func ScreenGetDefault() (*Screen, error) {
 	return toScreen(C.gdk_screen_get_default())
@@ -106,71 +94,11 @@ func (v *Screen) GetDisplay() (*Display, error) {
 	return toDisplay(C.gdk_screen_get_display(v.native()))
 }
 
-// GetNumber is a wrapper around gdk_screen_get_number().
-func (v *Screen) GetNumber() int {
-	return int(C.gdk_screen_get_number(v.native()))
-}
-
-// GetWidthMM is a wrapper around gdk_screen_get_width_mm().
-func (v *Screen) GetWidthMM() int {
-	return int(C.gdk_screen_get_width_mm(v.native()))
-}
-
-// GetHeightMM is a wrapper around gdk_screen_get_height_mm().
-func (v *Screen) GetHeightMM() int {
-	return int(C.gdk_screen_get_height_mm(v.native()))
-}
-
 func toString(c *C.gchar) (string, error) {
 	if c == nil {
 		return "", nilPtrErr
 	}
 	return C.GoString((*C.char)(c)), nil
-}
-
-// MakeDisplayName is a wrapper around gdk_screen_make_display_name().
-func (v *Screen) MakeDisplayName() (string, error) {
-	return toString(C.gdk_screen_make_display_name(v.native()))
-}
-
-// GetNMonitors is a wrapper around gdk_screen_get_n_monitors().
-func (v *Screen) GetNMonitors() int {
-	return int(C.gdk_screen_get_n_monitors(v.native()))
-}
-
-// GetPrimaryMonitor is a wrapper around gdk_screen_get_primary_monitor().
-func (v *Screen) GetPrimaryMonitor() int {
-	return int(C.gdk_screen_get_primary_monitor(v.native()))
-}
-
-// GetMonitorAtPoint is a wrapper around gdk_screen_get_monitor_at_point().
-func (v *Screen) GetMonitorAtPoint(x, y int) int {
-	return int(C.gdk_screen_get_monitor_at_point(v.native(), C.gint(x), C.gint(y)))
-}
-
-// GetMonitorAtWindow is a wrapper around gdk_screen_get_monitor_at_window().
-func (v *Screen) GetMonitorAtWindow(w *Window) int {
-	return int(C.gdk_screen_get_monitor_at_window(v.native(), w.native()))
-}
-
-// GetMonitorHeightMM is a wrapper around gdk_screen_get_monitor_height_mm().
-func (v *Screen) GetMonitorHeightMM(m int) int {
-	return int(C.gdk_screen_get_monitor_height_mm(v.native(), C.gint(m)))
-}
-
-// GetMonitorWidthMM is a wrapper around gdk_screen_get_monitor_width_mm().
-func (v *Screen) GetMonitorWidthMM(m int) int {
-	return int(C.gdk_screen_get_monitor_width_mm(v.native(), C.gint(m)))
-}
-
-// GetMonitorPlugName is a wrapper around gdk_screen_get_monitor_plug_name().
-func (v *Screen) GetMonitorPlugName(m int) (string, error) {
-	return toString(C.gdk_screen_get_monitor_plug_name(v.native(), C.gint(m)))
-}
-
-// GetMonitorScaleFactor is a wrapper around gdk_screen_get_monitor_scale_factor().
-func (v *Screen) GetMonitorScaleFactor(m int) int {
-	return int(C.gdk_screen_get_monitor_scale_factor(v.native(), C.gint(m)))
 }
 
 // GetResolution is a wrapper around gdk_screen_get_resolution().
@@ -181,11 +109,6 @@ func (v *Screen) GetResolution() float64 {
 // SetResolution is a wrapper around gdk_screen_set_resolution().
 func (v *Screen) SetResolution(r float64) {
 	C.gdk_screen_set_resolution(v.native(), C.gdouble(r))
-}
-
-// GetActiveWindow is a wrapper around gdk_screen_get_active_window().
-func (v *Screen) GetActiveWindow() (*Window, error) {
-	return toWindow(C.gdk_screen_get_active_window(v.native()))
 }
 
 // void 	gdk_screen_set_font_options ()
